+++ packages/components-vue/src/examples/index.ts
@@ -21,6 +21,8 @@ export { default as FlyoutMultilevelExampleCustomContent } from './FlyoutMultile
 export { default as HiddenAttributeExample } from './HiddenAttributeExample.vue';
 export { default as InlineNotificationExampleActionButton } from './InlineNotificationExampleActionButton.vue';
 export { default as InlineNotificationExampleEvents } from './InlineNotificationExampleEvents.vue';
+export { default as InputPasswordExample } from './InputPasswordExample.vue';
+export { default as InputPasswordExampleControlled } from './InputPasswordExampleControlled.vue';
 export { default as LinkTileProductExample } from './LinkTileProductExample.vue';
 export { default as ModalExampleAccessibility } from './ModalExample.vue';
 export { default as MultiSelectExample } from './MultiSelectExample.vue';